+++ Car/include/Car/internal/Vulkan/IndexBuffer.hpp
@@ -6,23 +6,23 @@
 namespace Car {
     class VulkanIndexBuffer : public IndexBuffer {
     public:
-        VulkanIndexBuffer(void* data, uint32_t size, Buffer::Usage usage, Buffer::Type type);
+        VulkanIndexBuffer(void* data, uint64_t size, Buffer::Usage usage, Buffer::Type type);
         virtual ~VulkanIndexBuffer() override;
 
         void releaseDeviceObjects();
 
         virtual void bind() const override;
 
-        virtual uint32_t getCount() const override { return mCount; }
-        virtual uint32_t getSize() const override { return mSize; }
+        virtual uint64_t getCount() const override { return mCount; }
+        virtual uint64_t getSize() const override { return mSize; }
         virtual Buffer::Usage getUsage() const override { return mUsage; }
         virtual Buffer::Type getType() const override { return mType; }
 
-        virtual void updateData(void* data, uint32_t size, uint32_t offset) override;
+        virtual void updateData(void* data, uint64_t size, uint64_t offset) override;
 
     private:
-        uint32_t mCount;
-        uint32_t mSize;
+        uint64_t mCount;
+        uint64_t mSize;
         Buffer::Usage mUsage;
         Buffer::Type mType;
