+++ Car/include/Car/internal/Vulkan/Renderer.hpp
@@ -11,7 +11,7 @@ namespace Car {
         virtual void InitImpl() override;
         virtual void ShutdownImpl() override;
         virtual void ClearColorImpl(float r, float g, float b, float a) override;
-        virtual void DrawCommandImpl(const Ref<VertexArray> va, uint32_t indicesCount) override;
+        virtual void DrawCommandImpl(const Ref<VertexArray> va, uint64_t indicesCount) override;
         virtual void SetViewportImpl(float x, float y, float width, float height, float minDepth,
                                      float maxDepth) override;
         virtual void SetScissorImpl(int32_t x, int32_t y, int32_t width, int32_t height) override;
